Add Params struct, which only uses X and Y as propertie for now Add Params constructor, that takes two int type values, x and y Add ParseArgs(string[] args) method (placeholder name)
ParseArgs method takes the array of strings args, which is passed to the
Main() from the command line. It's strings are analyzed through for cycle,
which uses a switch(case) to verify if any of them are equal to -x, for
example, and equals an int X to the next string in the array converted to
int, which is x's value given in the command line. This is repeated for
all arguments that are given and used to create the struct. The method
returns a new Params struct, using the values taken from the args array.